+++ src/main/java/api/GoblintClient.java
@@ -1,4 +1,4 @@
-package goblintclient;
+package api;
 
 /**
  * The Class {@link GoblintClient}.
