+++ src/main/java/api/json/GoblintMessageTypeAdapter.java
@@ -1,4 +1,4 @@
-package goblintclient;
+package api.json;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
