+++ src/main/java/api/messages/GoblintCFGResult.java
@@ -1,7 +1,7 @@
-package goblintclient.messages;
+package api.messages;
 
 /**
- * The Class GoblintCFG.
+ * The Class GoblintCFGResult.
  * <p>
  * Corresponding object to the Goblint cfg request response results in JSON.
  *
@@ -9,7 +9,7 @@ package goblintclient.messages;
  * @since 0.0.3
  */
 
-public class GoblintCFG {
+public class GoblintCFGResult {
     private String cfg;
 
     public String getCfg() {
