+++ src/main/java/api/messages/Params.java
@@ -1,4 +1,4 @@
-package goblintclient.messages;
+package api.messages;
 
 public class Params {
